feat: refactor mixin plugins and address PR comments
