+++ common/src/main/java/de/zannagh/armorhider/api/compat/readme.md
@@ -1,0 +1,4 @@
+Compat classes and resources for Armor Hider.
+
+None of the classes/files/... should carry any reference to other Armor Hider / mod-related code or Minecraft classes
+to prevent class load issues.
