+++ common/src/main/java/de/zannagh/armorhider/mixin/readme.md
@@ -1,0 +1,4 @@
+Mixin Plugins for Armor Hider (NeoForge + Fabric).
+
+None of the classes/files/... should carry any reference to other Armor Hider / mod-related code or Minecraft classes
+to prevent class load issues.
