+++ neoforge/src/main/java/de/zannagh/armorhider/client/MixinPlugin.java
@@ -1,118 +1,19 @@
 package de.zannagh.armorhider.client;
 
-import de.zannagh.armorhider.util.MixinUtil;
-import de.zannagh.armorhider.api.compat.CompatManager;
+import de.zannagh.armorhider.mixin.NeoForgeClientMixinPlugin;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.fml.loading.FMLEnvironment;
-import org.objectweb.asm.tree.ClassNode;
-import org.spongepowered.asm.mixin.extensibility.IMixinConfigPlugin;
-import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
+public class MixinPlugin extends NeoForgeClientMixinPlugin {
 
-public class MixinPlugin implements IMixinConfigPlugin {
-
-    private static final String PACKAGE = "de.zannagh.armorhider.client.mixin";
-
-    // Mixins listed unconditionally — Stonecutter source guards strip classes for
-    // incompatible versions, so getMixinClassesWherePresent filters them automatically.
-    private static final String[] MIXINS = new String[]{
-            // Always present
-            "PlayerMixin",
-            "GameRendererMixin",
-            "DevSkinMixin",
-            "TitleScreenSmokeMixin",
-            // All versions — Stonecutter guards per version range
-            "EntityRenderDispatcherMixin",
-            "LivingEntityMixin",
-            "LivingEntityRendererMixin",
-            "LivingEntityRenderStateMixin",
-            "lang.ClientLanguageMixin",
-            "hand.ItemEntityRendererMixin",
-            "hand.ItemInHandLayerMixin",
-            "hand.OffHandRenderMixin",
-            // Guarded by //? if >= 1.21.9 in source
-            "OptionsScreenMixin",
-            "SkinCustomizationScreenMixin",
-            "SkinCustomizationScreenLegacyMixin",
-            "bodyKneesAndToes.EquipmentRenderMixin",
-            // Transitional armor/elytra transparency for the 1.21.2/1.21.3 render-state era
-            // (EquipmentRenderMixin is >= 1.21.4). Stonecutter-gated; dropped where absent.
-            "bodyKneesAndToes.EquipmentLayerRendererLegacyMixin",
-            "hand.ItemRenderStateMixin",
-            "hand.SubmitNodeCollectorMixin",
-            "cape.ElytraRenderMixin",
-            "head.CustomHeadLayerMixin",
-            "head.SkullBlockRenderMixin",
-            // All versions — Stonecutter guards per version range
-            "bodyKneesAndToes.HumanoidArmorLayerMixin",
-            "bodyKneesAndToes.HumanoidArmorLayerRenderMixin",
-            "cape.CapeRenderMixin",
-            // Guarded by //? if < 1.21.9 in source
-            "hand.ModelPartMixin",
-            "hand.ItemRendererMixin",
-            "bodyKneesAndToes.NeoForgeArmorColorMixin",
-            "bodyKneesAndToes.NeoForgeHumanoidArmorLayerMixin",
-            "OptionsMixin",
-            "MinecraftClientMixin",
-            // Compat — @Pseudo, auto-skipped if target mod absent
-            "compat.wildfiregender.GenderArmorLayerMixin",
-            "compat.wildfiregender.GenderLegacyLayerMixin",
-            "compat.geckolib.GeckoLibArmorMixin",
-            "compat.waveycapes.WaveyCapesMixin",
-            // Compat - Pseudo and guarded in source for mekanism constant only (1.21.1 Neo)
-            "compat.mekanism.MekanismArmorMixin",
-            "compat.mekanism.MekaSuitArmorMixin",
-            "compat.emf.EmfModelPartMixin",
-            "compat.emf.EmfModelPartRootMixin",
-            "compat.curios.CuriosLayerMixin",
-            "compat.trinkets.TrinketRendererMixin",
-            "compat.accessories.AccessoriesRenderLayerMixin"
-    };
-    @Override
-    public boolean shouldApplyMixin(String targetClassName, String mixinClassName) {
-
-        return true;
-    }
-
-    @Override
-    public void acceptTargets(Set<String> myTargets, Set<String> otherTargets) {
-    }
-
-    @Override
-    public List<String> getMixins() {
+    public MixinPlugin() {
+        // FMLEnvironment is NeoForge-only, so the dedicated-server check lives here (not in the common
+        // intermediate) and is passed up as a plain boolean. Everything else — the mixin list,
+        // resource-probe onLoad, getMixins filtering — is inherited.
         //? if >= 1.21.9 {
-        /*if (FMLEnvironment.getDist() == Dist.DEDICATED_SERVER) {
-            return List.of();
-        }
+        /*super(FMLEnvironment.getDist() == Dist.DEDICATED_SERVER);
         *///?} else {
-
-        if (FMLEnvironment.dist == Dist.DEDICATED_SERVER) {
-            return List.of();
-        }
+        super(FMLEnvironment.dist == Dist.DEDICATED_SERVER);
         //?}
-        var mixinsToAdd = new ArrayList<>(List.of(MIXINS));
-        return MixinUtil.getMixinClassesWherePresent(PACKAGE, mixinsToAdd);
-    }
-
-    @Override
-    public void preApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {
-
-    }
-
-    @Override
-    public void postApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {
-    }
-
-    @Override
-    public void onLoad(String mixinPackage) {
-        CompatManager.setCompatFlagsByResourceProbing(MixinPlugin.class.getClassLoader());
-    }
-
-    @Override
-    public String getRefMapperConfig() {
-        return "";
     }
 }
